% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTARM_SupAndConf.R
\name{conf}
\alias{conf}
\title{conf}
\usage{
conf(x, y, DB)
}
\arguments{
\item{x}{The first of the two items to compute the confidence value for.}

\item{y}{The second of the two items to compute the confidence value for.}

\item{DB}{The data set the two items originate from.}
}
\value{
The confidence value.
}
\description{
Calculate the confidence value for a pair of items in the data.
}
\details{
The confidence value of items will need to be compared to the
minimum confidence value to determine which items to remove to accurately
generate the appropriate number of rules.
}
\author{
@skirwa
}
