% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTARM_InitializeRemove.R
\name{Initialize_Remove}
\alias{Initialize_Remove}
\title{Initialize_Remove}
\usage{
Initialize_Remove(DB, k, mconf)
}
\arguments{
\item{DB}{The data to generate rules for.}

\item{k}{The number of rules to be generated.}

\item{mconf}{The minimum accepted confidence value.}
}
\value{
The list of items to generate the desired number of rules.
}
\description{
Calculate the number of items to generate the desired number of rules,
and prepare to remove unneeded items.
}
\details{
This function is necessary at the start of the full FTARM algorithm to
cleanse the data of items useless to generating rules, and to determine the minimum
support value.
}
\author{
@skirwa
}
