% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTARM_SupAndConf.R
\name{minSup}
\alias{minSup}
\title{minSup}
\usage{
minSup(x, y, DB)
}
\arguments{
\item{x}{The first of the two items to compute the confidence value for.}

\item{y}{The second of the two items to compute the confidence value for.}

\item{DB}{The data set the two items originate from.}
}
\value{
The minimum acceptable support value for the pair of items.
}
\description{
Calculate the minimum support value for a pair of items in the data.
}
\details{
The minimum support value will be used for generating and expanding rules.
}
\author{
@skirwa
}
