% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTARM.R
\name{R}
\alias{R}
\title{R}
\usage{
R()
}
\description{
A NULL function to explain the package and provide references.
}
\details{
Since we are implementing an algorithm proposed by an article, we reference
the article here.
}
\references{
This package implements functions that are inspired by the
steps of an algoritm proposed in the following article:
"Fast Top-K association rule mining using generation property pruning"
Authors: Liu, Xiangyu' Niu, Xinzheng; Fournier-Viger, Philippe
Published 26 October 2020 by Springer Science+Business Media, LLC
}
\author{
@NessXToJason
}
