% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTARM_SortItem.R
\name{SortItems}
\alias{SortItems}
\title{SortItems}
\usage{
SortItems(DB)
}
\arguments{
\item{DB}{The data whose items will be sorted.}
}
\value{
A descending list of the items in the data based on their support values.
}
\description{
Sort the items in the data set based on their support value.
}
\details{
Updating the minimum support value in Initialize_Remove is a simpler process
when the data set is sorted by the items' support value.
}
\author{
@skirwa
}
