% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTARM_SupAndConf.R
\name{sup}
\alias{sup}
\title{sup}
\usage{
sup(x, DB)
}
\arguments{
\item{x}{The item whose support value will be calculated.}

\item{DB}{The data set the item originates from.}
}
\value{
The support value.
}
\description{
Calculate the support value for an item.
}
\details{
The support value of items will need to be compared to the
minimum support value to determine which items to remove to accurately
generate the appropriate number of rules.
}
\author{
@skirwa
}
