% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTARM_TIDFunctions.R
\name{tids}
\alias{tids}
\title{tids}
\usage{
tids(searchFor, DB)
}
\arguments{
\item{searchFor}{The element being searched for in each column of the database.}

\item{DB}{The database being searched for instances of searchFor.}
}
\value{
The number of instances of searchFor in DB (as an integer).
}
\description{
Find the number of transaction IDs that contain a certain value.
}
\details{
This function will be used as a helper function in the sup function.
}
\author{
@NessXToJason
}
