% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTARM_TIDFunctions.R
\name{tidsForTwo}
\alias{tidsForTwo}
\title{tidsForTwo}
\usage{
tidsForTwo(searchFor1, searchFor2, DB)
}
\arguments{
\item{searchFor1}{The first element being searched for in each column of the database.}

\item{searchFor2}{The second element being searched for in each column of the database.}

\item{DB}{The database being searched for instances of searchFor.}
}
\value{
The number of columns BOTH values appeared in.
}
\description{
Find the number of transaction IDs that contain BOTH of two given values.
}
\details{
Calculating confidence and minimum support/confidence values will require getting
the TIDs for multiple items.
}
\author{
@NessXToJason
}
